pyproject.toml
[tool.poetry]
name = "todays"
version = "0.1.0"
description = ""
authors = ["andreyPromaster <and45.rew@ya.ru>"]

[tool.poetry.dependencies]
python = "~=3.9"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.39"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.0"
email-validator = "^1.2.1"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-slugify = "^6.1.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
pre-commit = "^2.19.0"
pytest = "^7.1.2"
requests = "^2.28.1"
factory-boy = "^3.2.1"
SQLAlchemy-Utils = "^0.38.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
